requirements.txt
Flask==0.10.1
Flask-RESTful==0.3.4
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.10.4
aniso8601==1.1.0
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
decorator==4.0.4
ipython==4.0.0
ipython-genutils==0.1.0
iso8601==0.1.11
itsdangerous==0.24
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
python-dateutil==2.4.2
pytz==2015.7
recurly==2.2.17
requests==2.8.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
wsgiref==0.1.2
logbook
flask-sqlalchemy
marshmallow
